+++ json-formatter/CHANGELOG.md
@@ -19,6 +19,8 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Fixed
 
+* Process rules
+
 ## [2.2.0] - 2020-01-10
 
 ### Changed